v0.11

+++ src/main/java/com/xww/NewEngine/core/Component/impl/FpsComponent.java
@@ -9,7 +9,7 @@ import com.xww.NewEngine.gui.GameFrame;
 import java.awt.*;
 
 public class FpsComponent extends FreeComponent {
-    private int current_fps = GameFrame.context.current_fps;
+    private long current_fps = GameFrame.context.current_fps;
     public FpsComponent() {
         super(Vector.build(10, 10), GameFrame.PositionType.Screen, Vector.build(0, 0), AnchorMode.LeftTop, Vector.Zero(), Vector.build(0, 0), Integer.MAX_VALUE, -1);
         Timer fpsTimer = new Timer(1000, (obj)->{
